requirements.txt
scikit-image
pathlib
pillow